--- a/examples/BoundaryOp_timing/data/BOUT.inp
+++ b/examples/BoundaryOp_timing/data/BOUT.inp
@@ -1,0 +1,15 @@
+[mesh]
+nx = 128
+ny = 128
+nz = 128
+
+ixseps1 = -1
+
+[f_dirichlet]
+bndry_all = dirichlet
+
+[f_neumann]
+bndry_all = neumann
+
+[f_dirichlet_o3]
+bndry_all = dirichlet_o3