--- a/examples/BoundaryOp_timing/makefile
+++ b/examples/BoundaryOp_timing/makefile
@@ -1,0 +1,5 @@
+BOUT_TOP = ../..
+
+SOURCEC = BoundaryOp_timing.cxx
+
+include $(BOUT_TOP)/make.config