--- a/include/boundary_op.hxx
+++ b/include/boundary_op.hxx
@@ -8,37 +8,38 @@ class BoundaryModifier;
 #include "boundary_region.hxx"
 #include "field2d.hxx"
 #include "field3d.hxx"
+#include "unused.hxx"
 #include "vector2d.hxx"
 #include "vector3d.hxx"
-#include "unused.hxx"
 
 #include <cmath>
-#include <string>
 #include <list>
 #include <map>
+#include <string>
 using std::string;
 using std::list;
 
 /// An operation on a boundary
 class BoundaryOp {
 public:
-  BoundaryOp(bool apply_ddt = false) : bndry(nullptr), apply_to_ddt(apply_ddt),
-                                       val(0.), gen(nullptr), width(0) {}
+  BoundaryOp(bool apply_ddt = false)
+      : bndry(nullptr), apply_to_ddt(apply_ddt), val(0.), gen(nullptr), width(0) {}
   BoundaryOp(BoundaryRegion *region, int width_in = 0, bool apply_ddt = false)
-    : bndry(region), apply_to_ddt(apply_ddt), val(0.), gen(nullptr),
-      width(width_in ? width_in : region->width) {}
-  BoundaryOp(BoundaryRegion *region, BoutReal val_in, std::shared_ptr<FieldGenerator> g, int width_in = 0)
-    : bndry(region), apply_to_ddt(false), val(val_in), gen(std::move(g)),
-      width(width_in ? width_in : region->width) {}
+      : bndry(region), apply_to_ddt(apply_ddt), val(0.), gen(nullptr),
+        width(width_in ? width_in : region->width) {}
+  BoundaryOp(BoundaryRegion *region, BoutReal val_in, std::shared_ptr<FieldGenerator> g,
+             int width_in = 0)
+      : bndry(region), apply_to_ddt(false), val(val_in), gen(std::move(g)),
+        width(width_in ? width_in : region->width) {}
   virtual ~BoundaryOp() {}
 
   // Note: All methods must implement clone, except for modifiers (see below)
-  virtual BoundaryOp *clone(BoundaryRegion *UNUSED(region), const list<string> &UNUSED(args),
+  virtual BoundaryOp *clone(BoundaryRegion *UNUSED(region),
+                            const list<string> &UNUSED(args),
                             const std::map<std::string, std::string> &UNUSED(keywords)) {
     throw BoutException("BoundaryOp::clone not implemented");
     return nullptr;
   }
-
 
   /// Apply a boundary condition on field f
   virtual void apply(Field2D &f, BoutReal t = 0.) = 0;
@@ -59,76 +60,51 @@ public:
   /// Apply a boundary condition on ddt(f)
   virtual void apply_ddt(Field2D &f);
   virtual void apply_ddt(Field3D &f);
-  virtual void apply_ddt(Vector2D &f) {
-    apply(ddt(f));
-  }
-  virtual void apply_ddt(Vector3D &f) {
-    apply(ddt(f));
-  }
+  virtual void apply_ddt(Vector2D &f) { apply(ddt(f)); }
+  virtual void apply_ddt(Vector3D &f) { apply(ddt(f)); }
 
   BoundaryRegion *bndry;
-  const bool apply_to_ddt; // True if this boundary condition should be applied on the time derivatives, false if it should be applied to the field values
+  const bool apply_to_ddt; // True if this boundary condition should be applied on the
+                           // time derivatives, false if it should be applied to the field
+                           // values
 protected:
-  const BoutReal val; // constant value for boundary condition
+  const BoutReal val;                  // constant value for boundary condition
   std::shared_ptr<FieldGenerator> gen; // Generator
   const int width; // boundary width, stored in case we change it from the default
 };
 
 /// An operation on a boundary
-template<typename Derived, bool needs_delta = false>
+template <typename Derived, bool needs_delta = false>
 class BoundaryOpWithApply : public BoundaryOp {
 public:
   using BoundaryOp::BoundaryOp;
 
   // Note: All methods must implement clone, except for modifiers (see below)
-  virtual BoundaryOp* clone(BoundaryRegion *UNUSED(region), const list<string> &UNUSED(args)) {
+  virtual BoundaryOp *clone(BoundaryRegion *UNUSED(region),
+                            const list<string> &UNUSED(args)) {
     ASSERT1(false); // this implementation should never get called
     return nullptr;
   }
 
   /// Apply a boundary condition on field f
-  void apply(Field2D &f, BoutReal t = 0.) override {
-    applyTemplate(f, t);
-  }
-  void apply(Field3D &f, BoutReal t = 0.) override {
-    applyTemplate(f, t);
-  }
-
-protected:
-  //// Apply boundary condition at a point
-  //virtual void applyAtPoint(Field2D &UNUSED(f), BoutReal UNUSED(val), int
-  //    UNUSED(x), int UNUSED(bx), int UNUSED(y), int UNUSED(by), int UNUSED(z),
-  //    Coordinates* UNUSED(metric)) = 0;
-  //virtual void applyAtPoint(Field3D &UNUSED(f), BoutReal UNUSED(val), int
-  //    UNUSED(x), int UNUSED(bx), int UNUSED(y), int UNUSED(by), int UNUSED(z),
-  //    Coordinates* UNUSED(metric)) = 0;
-
-  //// Apply to staggered grid
-  //virtual void applyAtPointStaggered(Field2D &UNUSED(f), BoutReal UNUSED(val),
-  //    int UNUSED(x), int UNUSED(bx), int UNUSED(y), int UNUSED(by), int
-  //    UNUSED(z), Coordinates* UNUSED(metric)) = 0;
-  //virtual void applyAtPointStaggered(Field3D &UNUSED(f), BoutReal UNUSED(val),
-  //    int UNUSED(x), int UNUSED(bx), int UNUSED(y), int UNUSED(by), int
-  //    UNUSED(z), Coordinates* UNUSED(metric)) = 0;
-
-  //// extrapolate to further guard cells
-  //virtual void extrapolateFurther(Field2D &f, int x, int bx, int y, int by, int z) = 0;
-  //virtual void extrapolateFurther(Field3D &f, int x, int bx, int y, int by, int z) = 0;
+  void apply(Field2D &f, BoutReal t = 0.) override { applyTemplate(f, t); }
+  void apply(Field3D &f, BoutReal t = 0.) override { applyTemplate(f, t); }
 
 private:
-  template<typename T>
-  void applyTemplate(T &f, BoutReal t);
+  template <typename T> void applyTemplate(T &f, BoutReal t);
 };
 
 class BoundaryModifier : public BoundaryOp {
 public:
   BoundaryModifier(bool apply_ddt = false) : BoundaryOp(apply_ddt), op(nullptr) {}
   BoundaryModifier(BoundaryOp *operation, bool apply_ddt = false)
-    : BoundaryOp(operation->bndry, 0, apply_ddt), op(operation) {}
-  virtual BoundaryOp* cloneMod(BoundaryOp *op, const list<string> &args) = 0;
-  virtual BoundaryOpPar* cloneMod(BoundaryOpPar *UNUSED(op), const list<string> &UNUSED(args)) {
+      : BoundaryOp(operation->bndry, 0, apply_ddt), op(operation) {}
+  virtual BoundaryOp *cloneMod(BoundaryOp *op, const list<string> &args) = 0;
+  virtual BoundaryOpPar *cloneMod(BoundaryOpPar *UNUSED(op),
+                                  const list<string> &UNUSED(args)) {
     throw BoutException("BoundaryModifier should not be called on a BoundaryOpPar.");
   }
+
 protected:
   BoundaryOp *op;
 };