@ZedThree @JosephThomasParker @d7919, I made an example to check the performance of the `BoundaryOp`s. At least on my laptop, it seems that the refactored version was a lot slower than next. For 10000 calls to `f.applyBoundary()`:
`dirichlet` 1.05137 s (`next`) -> 4.63342 s (`BoundaryOp-refactor`)
`neumann` 3.37127  s (`next`) -> 6.1994 s (`BoundaryOp-refactor`)
`dirichlet_o3` 7.34798  s (`next`) -> 8.08621 s (`BoundaryOp-refactor`)

I've tried to jig things around to get more functions to inline and get back to the same performance. Getting rid of calls to `f.getNZ()` and `localmesh->GlobalZ(z)` seemed to help with `BoundaryDirichlet`. But `BoundaryDirichlet` implements its own `apply` method, so was pretty much the same as the old implementation anyway. `BoundaryNeumann` and `BoundaryDirichlet_O3` use the new base-class `apply()` and implement methods to apply the boundary condition at a single point that get called inside the loop. I thought using CRTP pattern to allow the `BoundaryOpWithApply<typename Derived>::applyTemplate`  method to use `static` methods of the `Derived` class so that they could be inlined, which helps a bit.

The thing that's confusing me is: adding `inline` keywords to some of the methods; in particular making `BoundaryNeumann::applyAtPoint(Field3D...)` `inline` reduced the execution time for the `neumann` boundary condition from about 2.7 s to 0.9s. However, adding `inline` to all the methods slows down both `dirichlet` and `neumann` (even though those methods haven't been changed). Is there some kind of limit to the number of methods that can be inlined in a file, and the compiler just stops inlining when the limit is reached?

The current state of this branch makes `examples/BoundaryOp_timing` run as fast as I've managed: adding more `inline` keywords slows it down (once I add a certain number of them).