@ZedThree 2018-11-06
`inline` is (mostly) just a suggestion to the compiler: "I'm confident that this can be inlined, please do so". The compiler normally has some heuristic on whether or not to actually do so; these include a limit on the function call depth to inline (it might not inline a 30 deep stack of function calls for instance). Another is on the size of the function(s), meaning it might not inline large functions even if you ask. I don't think there's a limit on the number of inline functions in a file though.

I believe inlining might also interact with the instruction cache as well, so inlining everything might plausibly have negative effects.