@johnomotani 2018-11-06
@ZedThree thanks for the quick reply. What you say confirms my vague ideas about what `inline` should do (I was surprised it had any effect at all). The part that really gets me is that adding `inline` to different boundary conditions (that aren't in the timing test) slows down the ones being timed. As an example, the extra `inlines` in 29757e40791981e886563dc9542db865e4c43b6a slow down 'dirichlet' and 'neumann' boundary conditions (on my desktop machine) from {0.67s, 0.69s} to {2.01s, 1.60s}. 'dirichlet_o3' isn't affected, it takes 2.16s before and after.

29757e40791981e886563dc9542db865e4c43b6a only added `inline` to `BoundaryDirichlet_O4::applyAtPoint`, `BoundaryDirichlet_smooth::applyAtPoint`, `BoundaryDirichlet_2ndOrder::applyAtPoint`, `BoundaryDirichlet_O5::applyAtPoint`, `BoundaryNeumann2::applyAtPoint`, `BoundaryNeumann_2ndOrder::applyAtPoint`, `BoundaryNeumann_O4::applyAtPoint`, `BoundaryNeumann_4thOrder::applyAtPoint`, and `BoundaryConstGradient::applyAtPoint`.
:confused: