@d7919 2018-11-06
Could it be just timing jitter?