@d7919 2018-11-06
I should also note that there are a few examples of performance tests with helpful macros in the `examples/performance/iterator*` directories which might help write your tests here.