@johnomotani 2018-11-06
>Could it be just timing jitter?

I don't think so, the timings vary by maybe +/-0.02s between repeated runs. The difference between the cases with/without more inlines is >1s.