@johnomotani 2018-11-06
The answer could also be that the way I've written the `BoundaryOp`s is horrible... pointers/suggestions for a different design welcome :+1: