@johnomotani 2018-11-07
Comparing to @dschwoerer's code in #1179, I realised that the main thing slowing down the boundary conditions was conditionals checking if the `FieldGenerator` is initialized inside the boundary condition loops. Moving the conditional outside the loops makes the new implementation significantly faster than the original. [Faster in most cases: oddly BoundaryDirichlet in the original implementation managed to be nearly as fast even with the `if`s inside the loops. Not sure how the compiler managed to optimize that, but it seems to have been a bit fragile, since with the old pattern changes elsewhere in the file slowed down BoundaryDirichlet].

There's now one branch with a `FieldGenerator` and one branch without, so a bit of code duplication, but it's only in two methods now, so seems worth it for the speed-up.