@johnomotani 2018-11-07
Parsing expressions is much slower, so I updated the boundary conditions to apply a constant value (in without using the expression parser.

I think this PR is ready to merge into #1334 now. Should I keep the `BoundaryOp_timing` example in and if so, (i) should I move it into `examples/performance`; (ii) does it need tidying up in the style of `examples/performance/iterator`, etc.?